+++ src/lists/list_mixins.lua
@@ -1,5 +1,5 @@
 local _, Addon = ...
-local Core = Addon.Core
+local Chat = Addon.Chat
 local E = Addon.Events
 local EventManager = Addon.EventManager
 local GetItemInfo = _G.GetItemInfo
@@ -26,7 +26,7 @@ function ListMixins:Add(itemID)
   if self._sv[itemID] then
     local itemLink = select(2, GetItemInfo(itemID))
     if itemLink then
-      Core:Print(L.ITEM_ALREADY_ON_LIST:format(itemLink, self.locale))
+      Chat:Print(L.ITEM_ALREADY_ON_LIST:format(itemLink, self.locale))
     end
   else
     self.toAdd[itemID] = true
@@ -44,7 +44,7 @@ function ListMixins:FinalizeAdd(item)
     -- Add to sv and print message if not loading from sv
     if not self._sv[item.ItemID] then
       self._sv[item.ItemID] = true
-      Core:Print(L.ADDED_ITEM_TO_LIST:format(item.ItemLink, self.locale))
+      Chat:Print(L.ADDED_ITEM_TO_LIST:format(item.ItemLink, self.locale))
     end
 
     -- Add item
@@ -57,7 +57,7 @@ function ListMixins:FinalizeAdd(item)
     return true
   end
 
-  Core:Print(reason)
+  Chat:Print(reason)
   return false
 end
 
@@ -74,7 +74,7 @@ function ListMixins:Remove(itemID, notify)
     local index = self:GetIndex(itemID)
     if index ~= -1 then
       local item = tremove(self.items, index)
-      Core:Print(
+      Chat:Print(
         L.REMOVED_ITEM_FROM_LIST:format(item.ItemLink, self.locale)
       )
       EventManager:Fire(E.ListItemRemoved, self, item)
@@ -82,7 +82,7 @@ function ListMixins:Remove(itemID, notify)
   elseif notify then
     local itemLink = select(2, GetItemInfo(itemID))
     if itemLink then
-      Core:Print(L.ITEM_NOT_ON_LIST:format(itemLink, self.locale))
+      Chat:Print(L.ITEM_NOT_ON_LIST:format(itemLink, self.locale))
     end
   end
 end
@@ -92,7 +92,7 @@ function ListMixins:RemoveAll()
   if next(self._sv) then
     for k in pairs(self._sv) do self._sv[k] = nil end
     for k in pairs(self.items) do self.items[k] = nil end
-    Core:Print(L.REMOVED_ALL_FROM_LIST:format(self.locale))
+    Chat:Print(L.REMOVED_ALL_FROM_LIST:format(self.locale))
     EventManager:Fire(E.ListRemovedAll, self)
   end
 end